+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - Backups will now backup your saves as well as some other files such as your game options
 - Scan certain Program Files folders when looking for Java versions on Windows
 - Fix Java warning prompts from showing when they shouldn't
+- Speed up launcher load time by hot loading the pack tabs